# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chainsmith API",
    "description": "Headless Rust API for the Chainsmith Flesh and Blood deck builder.",
    "license": {
      "name": "PolyForm-Noncommercial-1.0.0",
      "identifier": "PolyForm-Noncommercial-1.0.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/cards": {
      "get": {
        "operationId": "listCards",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "description": "Case-insensitive substring match on `name`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classes",
            "in": "query",
            "description": "Comma-separated list of class names. A card matches if it has any\nlisted class. Generic cards (no class) match only when the filter is\nomitted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "talents",
            "in": "query",
            "description": "Comma-separated list of talent names. A card matches if it has any\nlisted talent.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "types",
            "in": "query",
            "description": "Comma-separated list of card type names. A card matches if it has\nany listed type.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pitch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "costMin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "costMax",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Filter to cards that are currently `Legal` or `Restricted` in the\ngiven format (snapshot-based, not date-aware — for date-aware\nvalidation use `POST /validate`).",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FormatId"
            }
          },
          {
            "name": "legalForHero",
            "in": "query",
            "description": "Card id of a hero. Restricts results to cards whose supertypes are a\nsubset of the hero's effective supertypes (CR 1.1.3).",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CardId"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cards matching the filter, paginated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{id}": {
      "get": {
        "operationId": "getCard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Card unique id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card with all printings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDetail"
                }
              }
            }
          },
          "404": {
            "description": "Card not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/heroes": {
      "get": {
        "operationId": "listHeroes",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "description": "Case-insensitive substring match on `name`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classes",
            "in": "query",
            "description": "Comma-separated class names (any-match).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "talents",
            "in": "query",
            "description": "Comma-separated talent names (any-match).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "description": "`adult`, `young`, or `pit_fighter`. Multiple values not supported —\nmost queries want exactly one.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/HeroKind"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Filter to heroes currently `Legal` in the given format. CC-legal\nheroes that have been retired to Living Legend will *not* match\n`format=classic_constructed` (their `cc` legality is\n`living_legend_retired`); they will match `format=living_legend`.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FormatId"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Heroes matching the filter, paginated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeroListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Card": {
        "type": "object",
        "required": [
          "id",
          "name",
          "types",
          "subtypes",
          "classes",
          "talents",
          "keywords",
          "specializations",
          "legalitySummary"
        ],
        "properties": {
          "classes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Class"
            }
          },
          "cost": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "defense": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "flavorText": {
            "type": [
              "string",
              "null"
            ]
          },
          "functionalText": {
            "type": [
              "string",
              "null"
            ]
          },
          "hero": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HeroFacts"
              }
            ]
          },
          "id": {
            "$ref": "#/components/schemas/CardId"
          },
          "keywords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Keyword"
            }
          },
          "legalitySummary": {
            "$ref": "#/components/schemas/LegalitySummary"
          },
          "name": {
            "type": "string"
          },
          "pitch": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "power": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "specializations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HeroMoniker"
            }
          },
          "subtypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "talents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Talent"
            }
          },
          "typeText": {
            "type": [
              "string",
              "null"
            ]
          },
          "types": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardType"
            }
          },
          "weaponGrip": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WeaponGrip",
                "description": "Grip on weapon cards. `None` for non-weapons."
              }
            ]
          }
        }
      },
      "CardDetail": {
        "type": "object",
        "required": [
          "card",
          "printings"
        ],
        "properties": {
          "card": {
            "$ref": "#/components/schemas/Card"
          },
          "printings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Printing"
            }
          }
        }
      },
      "CardId": {
        "type": "string"
      },
      "CardListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardSummary"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque cursor; pass back as `cursor` to fetch the next page. `null`\nwhen there are no more results."
          }
        }
      },
      "CardSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "types",
          "classes",
          "talents",
          "legalitySummary"
        ],
        "properties": {
          "classes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Class"
            }
          },
          "cost": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "defaultPrinting": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingSummary",
                "description": "A representative printing for catalog tile rendering. The first one\nby collector number; `None` when the card has no printings."
              }
            ]
          },
          "defense": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "id": {
            "$ref": "#/components/schemas/CardId"
          },
          "legalitySummary": {
            "$ref": "#/components/schemas/LegalitySummary"
          },
          "name": {
            "type": "string"
          },
          "pitch": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "power": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "talents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Talent"
            }
          },
          "types": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardType"
            }
          }
        }
      },
      "CardType": {
        "type": "string",
        "enum": [
          "hero",
          "action",
          "attack_action",
          "attack_reaction",
          "defense_reaction",
          "instant",
          "equipment",
          "weapon",
          "mentor",
          "resource",
          "token"
        ]
      },
      "Class": {
        "type": "string",
        "enum": [
          "assassin",
          "bard",
          "brute",
          "guardian",
          "illusionist",
          "mechanologist",
          "merchant",
          "ninja",
          "ranger",
          "runeblade",
          "shapeshifter",
          "warrior",
          "wizard"
        ]
      },
...
      "Edition": {
        "type": "string",
        "enum": [
          "alpha",
          "unlimited",
          "first"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "Stable, machine-readable error codes. Clients are allowed to switch on\nthese. Adding a variant is additive; renaming or removing one is a\nbreaking change in production phase (see `.claude/rules/api-contract.md`).",
        "enum": [
          "database_error",
          "unsupported_format",
          "not_found",
          "unauthorized",
          "forbidden",
          "auth_not_configured"
        ]
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        }
      },
...
      "Foiling": {
        "type": "string",
        "enum": [
          "standard",
          "rainbow_foil",
          "cold_foil",
          "gold_foil"
        ]
      },
      "FormatId": {
        "type": "string",
        "enum": [
          "classic_constructed",
          "blitz",
          "commoner",
          "silver_age",
          "living_legend",
          "ultimate_pit_fight"
        ]
      },
...
      "HeroKind": {
        "type": "string",
        "enum": [
          "adult",
          "young",
          "pit_fighter"
        ]
      },
      "HeroListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HeroSummary"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "HeroMoniker": {
        "type": "string",
        "description": "The first comma-separated token of a hero's name (e.g. \"Bravo, Star of the\nShow\" -> \"Bravo\"). FaB specialization rules match on moniker, not full name."
      },
      "HeroSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "classes",
          "talents",
          "life",
          "essenceGrants",
          "legalitySummary"
        ],
        "properties": {
          "arcane": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "classes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Class"
            }
          },
          "defaultPrinting": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingSummary"
              }
            ]
          },
          "essenceGrants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EssenceGrant"
            }
          },
          "id": {
            "$ref": "#/components/schemas/CardId"
          },
          "intellect": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "kind": {
            "$ref": "#/components/schemas/HeroKind"
          },
          "legalitySummary": {
            "$ref": "#/components/schemas/LegalitySummary"
          },
          "life": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "talents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Talent"
            }
          }
        }
      },
      "Keyword": {
        "type": "string",
        "description": "Free-form keyword. Upstream prints new keywords every set, so an enum here\nwould be brittle and is not needed for legality checks (the validator does\nnot switch on keyword)."
      },
      "LegalitySummary": {
        "type": "object",
        "description": "Snapshot of per-format legality from the upstream `card.json` flat booleans.\nConvenience for catalog filtering only — the validator is the source of\ntruth for legality at a specific date.",
        "required": [
          "cc",
          "blitz",
          "commoner",
          "silverAge",
          "livingLegend",
          "upf"
        ],
        "properties": {
          "blitz": {
            "$ref": "#/components/schemas/FormatStatus"
          },
          "cc": {
            "$ref": "#/components/schemas/FormatStatus"
          },
          "commoner": {
            "$ref": "#/components/schemas/FormatStatus"
          },
          "livingLegend": {
            "$ref": "#/components/schemas/FormatStatus"
          },
          "silverAge": {
            "$ref": "#/components/schemas/FormatStatus"
          },
          "upf": {
            "$ref": "#/components/schemas/FormatStatus"
          }
        }
      },
...
      "Printing": {
        "type": "object",
        "required": [
          "id",
          "cardId",
          "set",
          "setReleaseDate",
          "edition",
          "foiling",
          "treatment",
          "rarity",
          "collectorNumber"
        ],
        "properties": {
          "artist": {
            "type": [
              "string",
              "null"
            ]
          },
          "cardId": {
            "$ref": "#/components/schemas/CardId"
          },
          "collectorNumber": {
            "type": "string"
          },
          "edition": {
            "$ref": "#/components/schemas/Edition"
          },
          "foiling": {
            "$ref": "#/components/schemas/Foiling"
          },
          "id": {
            "$ref": "#/components/schemas/PrintingId"
          },
          "imageUrl": {
            "type": [
              "string",
              "null"
            ],
            "description": "CDN URL for the card image, sourced from upstream. Clients (web,\niOS, Android) load this URL directly from the LSS CDN — we do not\nproxy or cache imagery on our infrastructure (see\n`.claude/rules/security.md`)."
          },
          "rarity": {
            "$ref": "#/components/schemas/Rarity"
          },
          "set": {
            "$ref": "#/components/schemas/SetCode"
          },
          "setReleaseDate": {
            "type": "string",
            "format": "date"
          },
          "treatment": {
            "$ref": "#/components/schemas/Treatment"
          }
        }
      },
      "PrintingId": {
        "type": "string"
      },
      "PrintingSummary": {
        "type": "object",
        "required": [
          "id",
          "set",
          "rarity"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/PrintingId"
          },
          "imageUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "rarity": {
            "$ref": "#/components/schemas/Rarity"
          },
          "set": {
            "$ref": "#/components/schemas/SetCode"
          }
        }
      },
      "Rarity": {
        "type": "string",
        "enum": [
          "common",
          "rare",
          "super_rare",
          "majestic",
          "legendary",
          "fabled",
          "promo",
          "token"
        ]
      },
...
      "Talent": {
        "type": "string",
        "enum": [
          "draconic",
          "earth",
          "elemental",
          "ice",
          "light",
          "lightning",
          "shadow",
          "wind"
        ]
      },
...
      "WeaponGrip": {
        "type": "string",
        "enum": [
          "one_handed",
          "two_handed"
        ]
      }
    }
  }
}